Started breaking up CWH into visible components (Main, Dashboard, Printers, Printer, EditPrinter) Added Font Awesome resources Added "started" property for Printer so for a pretty gui Started refactoring for PrinterService API Depricated methods in MachineService API Screen shots of new GUI
