+++ host/resourcesnew/cardFooter.html
@@ -1,0 +1,2 @@
+<small class="pull-left">By Wes Gilster</small>
+<a class="btn btn-link btn-xs pull-right" href="https://github.com/area515/Creation-Workshop-Host/blob/master/README.md">Help</a>
