--- host/resourcesnew/test.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<!-- saved from url=(0026)http://demo.bootcards.org/ -->
-<html lang="en"><head><meta http-equiv="Content-Type" content="text/html; charset=UTF-8" ng-app="cwhApp">
-  <meta charset="utf-8">
-  <!--<link rel="shortcut icon" href="../../docs-assets/ico/favicon.png">-->
-
-    <meta name="mobile-web-app-capable" content="yes">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0, user-scalable=no">
-
-	<script src="./angular/js/angular.min.js"></script> 
-	<script src="./cwh/js/main.js"></script> 
-</head>
-<body id="bootcards">
-{{dashboard.printers.name}}
-</body></html>
