+++ host/src/org/area515/resinprinter/services/MachineService.java
@@ -52,11 +52,13 @@ public class MachineService {
 	
 	private MachineService(){}
 	
+	 @Deprecated
 	 @GET
 	 @Path("printers")
 	 @Produces(MediaType.APPLICATION_JSON)
 	// @RolesAllowed("Admin")
 	 public List<String> getPrinters() {
+		 
 		 List<PrinterConfiguration> identifiers = HostProperties.Instance().getPrinterConfigurations();
 		 List<String> identifierStrings = new ArrayList<String>();
 		 for (PrinterConfiguration current : identifiers) {
@@ -66,6 +68,7 @@ public class MachineService {
 		 return identifierStrings;
 	 }
 	 
+	 @Deprecated
 	 @GET
 	 @Path("ports")
 	 @Produces(MediaType.APPLICATION_JSON)
@@ -80,9 +83,17 @@ public class MachineService {
 	 }
 	 
 	 @GET
+	 @Path("serialPorts/list")
+	 @Produces(MediaType.APPLICATION_JSON)
+	 public List<String> getSerialPorts() {
+		 return getPorts();
+	 }
+	 
+	 @Deprecated
+	 @GET
 	 @Path("displays")
 	 @Produces(MediaType.APPLICATION_JSON)
-	 public List<String> getDisplays() {
+	 public List<String> getDisplaysOld() {
 		 List<GraphicsDevice> devices = DisplayManager.Instance().getDisplayDevices();
 		 List<String> deviceStrings = new ArrayList<String>();
 		 for (GraphicsDevice current : devices) {
@@ -92,163 +103,23 @@ public class MachineService {
 		 return deviceStrings;
 	 }
 	 
-	 
-	 
-	 
-	 
-	 
-	 
-	 
-	 
-	 
-	 
-	 //These are some help methods:
-	 public static void main(String[] args) {
-		 new MachineService().createDefaultPrintersInMachinesDirectory();
-	 }
-	 
-	 private void createDefaultPrintersInMachinesDirectory() {
-		 PrinterConfiguration configuration = createTemplatePrinter(
-				 "mUVe 1 DLP (Testing)", 
-				 DisplayManager.SIMULATED_DISPLAY, 
-				 ConsoleCommPort.CONSOLE_COMM_PORT, 
-				 134, 75, 185);
-		 configuration.getSlicingProfile().getSelectedInkConfig().setNumberOfFirstLayers(10);
-		 configuration.getSlicingProfile().getSelectedInkConfig().setFirstLayerExposureTime(20000);
-		 configuration.getSlicingProfile().getSelectedInkConfig().setExposureTime(8000);
-		 configuration.getSlicingProfile().setgCodeHeader(
-				 "G21 ;Set units to be mm\n" +
-				 "G91 ;Relative Positioning\n" +
-				 "G28 ; Home Printer\n" +
-				 "M650 D$ZLiftDist S$ZLiftRate P0;mUVe 1 Prefs\n" +
-				 "M17 ;Enable motors");
-		 configuration.getSlicingProfile().setgCodeFooter(
-				 "M18 ;Disable Motors");
-		 configuration.getSlicingProfile().setgCodeLift(
-				 "M651; Do mUVe 1 Peel Move\n" + 
-				 "G1 Z${((LayerThickness) * ZDir)}");
-		 configuration.getSlicingProfile().setZLiftDistanceGCode("M650 D${ZLiftDist} S${ZLiftRate}");
-		 configuration.getSlicingProfile().setZLiftSpeedGCode("M650 D${ZLiftDist} S${ZLiftRate}");
-		 configuration.getSlicingProfile().setzLiftSpeedCalculator("var value = 0.25;\n"
-		 		+ "if ($CURSLICE > $NumFirstLayers) {\n"
-		 		+ " value = 4.6666666666666705e+000 * Math.pow($buildAreaMM,0) + -7.0000000000000184e-003 * Math.pow($buildAreaMM,1) + 3.3333333333333490e-006 * Math.pow($buildAreaMM,2);\n"
-		 		+ "}\n"
-		 		+ "value");
-		 configuration.getSlicingProfile().setzLiftDistanceCalculator("var value = 9.0;\n"
-		 		+ "if ($CURSLICE > $NumFirstLayers) {\n"
-			 	+ " value = 3.5555555555555420e+000 * Math.pow($buildAreaMM,0) + 4.3333333333334060e-003 * Math.pow($buildAreaMM,1) + 1.1111111111110492e-006 * Math.pow($buildAreaMM,2);\n"
-			 	+ "}\n"
-			 	+ "value");
-		 configuration.getSlicingProfile().setExposureTimeCalculator("var value = $FirstLayerTime;\n"
-		 		+ "if ($CURSLICE > $NumFirstLayers) {\n"
-		 		+ "	value = $LayerTime\n"
-		 		+ "}\n"
-			 	+ "value");
-		 configuration.getSlicingProfile().setProjectorGradientCalculator(			    
-		 		"function getFractions(count, start, end) {\n" + 
-				"	var incrementAmount = (end - start) / count;\n" +
-				"	var fractions = [];\n" + 
-				"	for (t = 0; t < count; t++) {\n" +
-				"		fractions[t] = start + incrementAmount * t;\n" +
-				"	}\n" +
-				"	//return new float[]{0, 1};\n" +
-				"	return fractions;\n" + 
-	 			"}\n" +
-	 			"function getColors(fractions, start, stop) {\n" + 
-	 			"	var colors = [];\n" +
-	 			"	var colorRange = stop - start;\n" + 
-	 			"	var atanDivergencePoint = Math.PI / 2;\n" +
-	 			"	for (t = 0; t < fractions.length; t++) {\n" +
-	 			"		colors[t] = new Packages.java.awt.Color(0, 0, 0, (java.lang.Integer)(Math.atan(fractions[t] * atanDivergencePoint) * colorRange + start));\n" +
-				"	}\n" + 
-				"	//return new Packages.java.awt.Color[]{new Packages.java.awt.Color(0, 0, 0, (java.lang.Integer)(opacityLevelModel.getValue()/(float)opacityLevelModel.getMaximum())), new Packages.java.awt.Color(0, 0, 0, 0)};\n" +
-				"	return colors;\n" + 
-	 			"}\n" +
-	 			"var bulbCenter = new Packages.java.awt.geom.Point2D.Double($buildPlatformXPixels / 2, $buildPlatformYPixels / 2);\n" +
-	 			"var bulbFocus = new Packages.java.awt.geom.Point2D.Double($buildPlatformXPixels / 2, $buildPlatformYPixels / 2);\n" +
-	 			"var totalSizeOfGradient = $buildPlatformXPixels > $buildPlatformYPixels?$buildPlatformXPixels:$buildPlatformYPixels;\n" +
-	 			"var fractions = getFractions(totalSizeOfGradient, 0, 1);\n" +
-	 			"var colors = getColors(fractions, 0.2, 0);//Let's start with 20% opaque in the center of the projector bulb\n" +
-	 			"new Packages.java.awt.RadialGradientPaint(\n" +
-	 			"	bulbCenter,\n" + 
-	 			"	totalSizeOfGradient,\n" +
-				"	bulbFocus,\n" +
-				"	fractions,\n" + 
-				"	colors,\n" +
-				"	MultipleGradientPaint.CycleMethod.NO_CYCLE)");
-		 
-		try {
-			HostProperties.Instance().addPrinterConfiguration(configuration);
-		} catch (AlreadyAssignedException e) {
-			e.printStackTrace();
-		}
+	 @GET
+	 @Path("graphicsDisplays/list")
+	 @Produces(MediaType.APPLICATION_JSON)
+	 public List<String> getDisplays() {
+		 return getDisplaysOld();
 	 }
 	 
 	 
-	 private PrinterConfiguration createTemplatePrinter(String printername, String displayId, String comport, double physicalProjectionMMX, double physicalProjectionMMY, double buildHeightMMZ) {
-			PrinterConfiguration currentConfiguration = new PrinterConfiguration(printername, printername);
-			ComPortSettings settings = new ComPortSettings();
-			settings.setPortName(comport);
-			settings.setDatabits(8);
-			settings.setHandshake("None");
-			settings.setStopbits("One");
-			settings.setParity("None");
-			settings.setSpeed(115200);
-			MotorsDriverConfig motors = new MotorsDriverConfig();
-			motors.setComPortSettings(settings);
-			MonitorDriverConfig monitor = new MonitorDriverConfig();
-			
-			MachineConfig machineConfig = new MachineConfig();
-			machineConfig.setMotorsDriverConfig(motors);
-			machineConfig.setMonitorDriverConfig(monitor);
-			machineConfig.setOSMonitorID(displayId);
-			machineConfig.setName(printername);
-			machineConfig.setPlatformXSize(physicalProjectionMMX);
-			machineConfig.setPlatformYSize(physicalProjectionMMY);
-			machineConfig.setPlatformZSize(buildHeightMMZ);
-			
-			SlicingProfile slicingProfile = new SlicingProfile();
-			slicingProfile.setLiftDistance(5.0);
-			slicingProfile.setLiftFeedRate(50);
-			slicingProfile.setDirection(BuildDirection.Bottom_Up);
-			try {
-				GraphicsDevice device = DisplayManager.Instance().getDisplayDevice(DisplayManager.LAST_AVAILABLE_DISPLAY);
-				monitor.setDLP_X_Res(device.getDefaultConfiguration().getBounds().getWidth());
-				monitor.setDLP_Y_Res(device.getDefaultConfiguration().getBounds().getHeight());
-				machineConfig.setxRenderSize((int)monitor.getDLP_X_Res());
-				machineConfig.setyRenderSize((int)monitor.getDLP_Y_Res());
-				slicingProfile.setxResolution((int)monitor.getDLP_X_Res());
-				slicingProfile.setyResolution((int)monitor.getDLP_Y_Res());
-			} catch (InappropriateDeviceException e) {
-				e.printStackTrace();
-				throw new IllegalArgumentException("Couldn't get screen device");
-			}
-			
-			InkConfig ink = new InkConfig();
-			ink.setName("Default");
-			ink.setNumberOfFirstLayers(3);
-			ink.setResinPriceL(65.0);
-			ink.setSliceHeight(0.1);
-			ink.setFirstLayerExposureTime(5000);
-			ink.setExposureTime(1000);
-			
-			List<InkConfig> configs = new ArrayList<InkConfig>();
-			configs.add(ink);
-			
-			slicingProfile.setInkConfigs(configs);
-			slicingProfile.setSelectedInkConfigName("Default");
-			slicingProfile.setDotsPermmX(monitor.getDLP_X_Res() / physicalProjectionMMX);
-			slicingProfile.setDotsPermmY(monitor.getDLP_Y_Res() / physicalProjectionMMY);
-			slicingProfile.setFlipX(false);
-			slicingProfile.setFlipY(true);
-			
-			currentConfiguration.setSlicingProfile(slicingProfile);
-			currentConfiguration.setMachineConfig(machineConfig);
-			
-			currentConfiguration.setName(printername);
-			return currentConfiguration;
-	 }
 	 
+	 
+	 
+	 
+	 
+	 
+	 
+	 
+
 	 
 	 
 	 
@@ -260,14 +131,14 @@ public class MachineService {
 	 
 	 //The following methods are for Printers
 	 //======================================
-
+	 @Deprecated
 	 @GET
 	 @Path("createprinter/{printername}/{display}/{comport}")
 	 @Produces(MediaType.APPLICATION_JSON)
 	 public MachineResponse createPrinter(@PathParam("printername") String printername, @PathParam("display") String displayId, @PathParam("comport") String comport) {
 		//TODO: This data needs to be set by the user interface...
 		//========================================================
-		PrinterConfiguration currentConfiguration = createTemplatePrinter(printername, displayId, comport, 134, 75, 185);
+		PrinterConfiguration currentConfiguration = PrinterService.INSTANCE.createTemplatePrinter(printername, displayId, comport, 134, 75, 185);
 		if (displayId.equals(DisplayManager.SIMULATED_DISPLAY) &&
 			comport.equals(ConsoleCommPort.CONSOLE_COMM_PORT)) {
 			currentConfiguration.getSlicingProfile().setgCodeLift("Lift Z; Lift the platform");
@@ -285,6 +156,7 @@ public class MachineService {
 		}
 	 }
 	 
+	 @Deprecated
 	 @GET
 	 @Path("deleteprinter/{printername}")
 	 @Produces(MediaType.APPLICATION_JSON)
